Move NFTExt to pallet_common
