+++ pallet/token-approvals/src/lib.rs
@@ -25,8 +25,7 @@
 
 use frame_support::pallet_prelude::*;
 use frame_system::pallet_prelude::*;
-use pallet_nft::traits::NFTExt;
-use seed_pallet_common::OnTransferSubscriber;
+use seed_pallet_common::{NFTExt, OnTransferSubscriber};
 use seed_primitives::{AssetId, Balance, CollectionUuid, TokenId};
 use sp_runtime::{traits::Zero, DispatchResult};
